core/security.py
import bcrypt
from datetime import datetime, timedelta, timezone
from jose import jwt
from core.config import settings

def get_password_hash(password: str) -> str:
    # Генерируем соль и хешируем (bcrypt сам ограничит до 72 байт без ошибок)
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(password.encode('utf-8'), salt)
    return hashed.decode('utf-8')

def verify_password(plain_password: str, hashed_password: str) -> bool:
    try:
        return bcrypt.checkpw(
            plain_password.encode('utf-8'), 
            hashed_password.encode('utf-8')
        )
    except Exception:
        return False

def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=15))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)


def create_refresh_token(data: dict):
    expires_delta = timedelta(days=30)
    return create_access_token(data, expires_delta)